Move the install scritps and util from OpenWhisk to OpenWhisk-catalog
This patch is the first step to move the calalog from OpenWhisk
to OpenWhisk-catalog, consisting of the movement of all the install
scripts and util under openwhisk/catalog into
openwhisk-catalog/packages.

An additional folder named packages-test is added as well. Please
refer to the README.md under it for more information on the tests.

Closes-Bug: #4