Add the automation of Travis CI
This patch will add the travis ci to test the catalog installation
on the openwhisk environment, and check if each package works or not.
It is supposed to make sure each new pull request for this repo
works good in order to merge.

Security keys to test each package should be resolved here. If not,
another PR is welcome to open.

Partial-Closes-Bug: #24